changing type of data in dataProvider to float

+++ src/test/java/org/academy/kata/EightTest.java
@@ -19,9 +19,9 @@ public class EightTest extends EightDataProvider {
 
 
     @Test(dataProvider = "kmPerLiterProvider")
-    public void testMpgToKPM(double mpg, double expectedResult, IEight iEight) {
-        float actualResult = iEight.mpgToKPM((float)mpg);
-        Assert.assertEquals(actualResult, (float)expectedResult, 0.001);
+    public void testMpgToKPM(float mpg, float expectedResult, IEight iEight) {
+        float actualResult = iEight.mpgToKPM(mpg);
+        Assert.assertEquals(actualResult, expectedResult, 0.001);
     }
 
     @Test
